requirements.txt
ollama
streamlit
faiss-cpu
PyPDF2
sentence-transformers
transformers
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
langchain
langchain-community
langchain-ollama
langchain_huggingface
sentencepiece 